requirements.txt
flask==3.1.0
flask-cors==5.0.1
python-jobspy==1.1.75
gunicorn==23.0.0
